[DIDA] refactor. NFT만들기 > NFT만들기관련 클래스 네이밍 변경
