--- presentation/src/main/java/com/dida/android/presentation/views/nav/add/addnftprice/AddNftPriceActionHandler.kt
@@ -1,5 +1,0 @@
-package com.dida.android.presentation.views.nav.add.addnftprice
-
-interface AddNftPriceActionHandler {
-    fun onOkBtnClicked()
-}
