--- presentation/src/main/java/com/dida/android/presentation/views/nav/add/addnftprice/AddNftPriceNavigationAction.kt
@@ -1,5 +1,0 @@
-package com.dida.android.presentation.views.nav.add.addnftprice
-
-sealed class AddNftPriceNavigationAction {
-    object NavigateToDismiss: AddNftPriceNavigationAction()
-}
