+++ presentation/src/main/java/com/dida/android/presentation/views/nav/add/bottom/AddKeepNftBottomSheet.kt
@@ -1,22 +1,22 @@
-package com.dida.android.presentation.views.nav.add
+package com.dida.android.presentation.views.nav.add.bottom
 
 import androidx.fragment.app.viewModels
 import com.dida.android.R
-import com.dida.android.databinding.BottomAddNftBinding
+import com.dida.android.databinding.BottomAddKeepNftBinding
 import com.dida.common.base.BaseBottomSheetDialogFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class AddNftBottomSheet(
+class AddKeepNftBottomSheet(
     val itemClick: () -> Unit
-) : BaseBottomSheetDialogFragment<BottomAddNftBinding, AddNftViewModel>() {
+) : BaseBottomSheetDialogFragment<BottomAddKeepNftBinding, AddKeepViewModel>() {
 
     private val TAG = "AddNftBottomSheet"
 
     override val layoutResourceId: Int
-        get() = R.layout.bottom_add_nft
+        get() = R.layout.bottom_add_keep_nft
 
-    override val viewModel: AddNftViewModel by viewModels()
+    override val viewModel: AddKeepViewModel by viewModels()
 
     override fun initStartView() {
         binding.vm = viewModel
