+++ presentation/src/main/java/com/dida/android/presentation/views/nav/add/bottom/AddKeepViewModel.kt
@@ -1,10 +1,10 @@
-package com.dida.android.presentation.views.nav.add
+package com.dida.android.presentation.views.nav.add.bottom
 
 import com.dida.common.base.BaseViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class AddNftViewModel @Inject constructor() : BaseViewModel() {
+class AddKeepViewModel @Inject constructor() : BaseViewModel() {
     private val TAG = "AddNftViewModel"
 }
