+++ presentation/src/main/java/com/dida/android/presentation/views/nav/add/bottom/AddSaleNftActionHandler.kt
@@ -1,0 +1,5 @@
+package com.dida.android.presentation.views.nav.add.bottom
+
+interface AddSaleNftActionHandler {
+    fun onOkBtnClicked()
+}
