+++ presentation/src/main/java/com/dida/android/presentation/views/nav/add/bottom/AddSaleNftNavigationAction.kt
@@ -1,0 +1,5 @@
+package com.dida.android.presentation.views.nav.add.bottom
+
+sealed class AddSaleNftNavigationAction {
+    object NavigateToDismiss: AddSaleNftNavigationAction()
+}
