+++ presentation/src/main/java/com/dida/android/presentation/views/nav/add/purpose/AddPurposeActionHandler.kt
@@ -1,4 +1,4 @@
-package com.dida.android.presentation.views.nav.add.addpurpose
+package com.dida.android.presentation.views.nav.add.purpose
 
 interface AddPurposeActionHandler {
     fun onTypeNotSaleClicked()
