+++ presentation/src/main/java/com/dida/android/presentation/views/nav/add/purpose/AddPurposeNavigationAction.kt
@@ -1,4 +1,4 @@
-package com.dida.android.presentation.views.nav.add.addpurpose
+package com.dida.android.presentation.views.nav.add.purpose
 
 sealed class AddPurposeNavigationAction {
     object NavigateToSaled: AddPurposeNavigationAction()
