+++ presentation/src/main/java/com/dida/android/presentation/views/nav/add/purpose/AddPurposeViewModel.kt
@@ -1,4 +1,4 @@
-package com.dida.android.presentation.views.nav.add.addpurpose
+package com.dida.android.presentation.views.nav.add.purpose
 
 import com.dida.common.base.BaseViewModel
 import com.dida.domain.flatMap
