+++ presentation/src/main/res/layout/bottom_add_keep_nft.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="vm"
-            type="com.dida.android.presentation.views.nav.add.AddNftViewModel" />
+            type="com.dida.android.presentation.views.nav.add.bottom.AddKeepViewModel" />
     </data>
 
     <FrameLayout
