+++ presentation/src/main/res/layout/bottom_add_sale_nft.xml
@@ -4,7 +4,7 @@
     <data>
         <variable
             name="vm"
-            type="com.dida.android.presentation.views.nav.add.addnftprice.AddNftPriceViewModel" />
+            type="com.dida.android.presentation.views.nav.add.bottom.AddSaleNftViewModel" />
     </data>
 
     <FrameLayout
