+++ presentation/src/main/res/layout/fragment_add.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="vm"
-            type="com.dida.android.presentation.views.nav.add.addmain.AddViewModel" />
+            type="com.dida.android.presentation.views.nav.add.main.AddViewModel" />
     </data>
